+++ include/kImageAnnotator/KImageAnnotator.h
@@ -47,7 +47,7 @@ public:
 	void showAnnotator();
 	void showCropper();
 	void showScaler();
-	void setSettingsVisible(bool isVisible);
+	void setSettingsCollapsed(bool isCollapsed);
 
 public Q_SLOTS:
 	void loadImage(const QPixmap &pixmap);